* strmisc: 	WideCharConverter: all converting methods are const methods  * commom.h and deepthught.cpp 	DocIds are uint32_t now.

+++ src/common.h
@@ -8,7 +8,7 @@
 #include <set>
 #include <utility>
 
-typedef u_int64_t docid_t;
+typedef uint32_t docid_t;
 
 //!A simple struct-like class just to store information about a page.
 typedef std::pair<std::string, docid_t> PageRef;
